+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Add `card` model and associated methods (see: `lib/graphql/card.ts`)
+- Add a method to categorize transactions (`models.transaction.categorize`)
+- Add a method to fetch wire transfer suggestions (`models.transfer.suggestions`)
 
 ## [0.22.2] - 2019-11-26
 ### Fixed